Create LangTest.php
