+++ tests/NumberTest.php
@@ -6,7 +6,6 @@ namespace Serhii\Tests;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use Serhii\ShortNumber\Exceptions\LargeNumberException;
-use Serhii\ShortNumber\Lang;
 use Serhii\ShortNumber\Number;
 
 final class NumberTest extends TestCase
@@ -66,26 +65,5 @@ final class NumberTest extends TestCase
     {
         $this->expectException(LargeNumberException::class);
         Number::short(1_000_000_000_000_000_000);
-    }
-
-    /**
-     * @param array<string,string> $overwrites
-     */
-    #[DataProvider('providerForOverwrites')]
-    public function testOverwrites(string $expect, int $number, array $overwrites): void
-    {
-        Lang::set('en', overwrites: $overwrites);
-        $this->assertSame($expect, Number::short($number));
-    }
-
-    public static function providerForOverwrites(): array
-    {
-        return [
-            ['1kilo', 1000, ['k' => 'kilo']],
-            ['2', 2000, ['k' => '']],
-            ['33k', 33000, ['k' => 'k']],
-            ['A kilo', 1000, ['1k' => 'A kilo']],
-            ['123 kilo', 123456, ['k' => ' kilo']],
-        ];
     }
 }
